requirements.txt
# Core ML/AI Dependencies
torch>=2.0.0 --index-url https://download.pytorch.org/whl/cpu
torchvision>=0.15.0 --index-url https://download.pytorch.org/whl/cpu
torchaudio>=2.0.0 --index-url https://download.pytorch.org/whl/cpu
transformers>=4.36.0
accelerate>=0.27.0
huggingface_hub>=0.19.0
diffusers>=0.25.0

# Audio Processing
gTTS>=2.2.3
mutagen>=1.45.1
pydub>=0.25.1
scipy==1.10.1

# Image and Video Processing
pillow>=8.3.0
moviepy>=1.0.3

# Google API Integration
google-auth-oauthlib>=0.4.6
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0

# HTTP and Networking
aiohttp>=3.8.0
openai>=1.3.0

# Utilities
python-dotenv>=0.19.0